requirement.txt
fastapi==0.127.0
uvicorn==0.40.0
requests==2.32.5
python-dotenv==1.2.1
pydantic==2.12.5
numpy==2.3.3
pandas==2.3.3